refactor(api): clarify wrangler config helper
Rename the local database URL helper away from the overloaded env.ts name and remove the stale env.d.ts tsconfig include. The Cloudflare env augmentation stays explicit in api-public-origin.d.ts because it documents the dev-only origin override outside generated Wrangler types.
