+++ apps/api/tsconfig.json
@@ -5,5 +5,5 @@
 		"jsxImportSource": "hono/jsx",
 		"types": ["bun", "./worker-configuration.d.ts"]
 	},
-	"include": ["worker", "scripts", "env.d.ts", "api-public-origin.d.ts"]
+	"include": ["worker", "scripts", "api-public-origin.d.ts"]
 }
